requirements.txt
alembic>=1.7.0
protobuf>=4.24.0
grpcio-reflection>=1.59.0 
grpcio==1.59.0
grpcio-tools==1.59.0
sqlalchemy>=1.4.23
